support option, box on no_std
- add test examples for box option features on no_std
