+++ lib/Cargo.toml
@@ -42,3 +42,4 @@ nesting = [
 ]
 none_as_default = ["option"]
 keep_none = ["option"]
+
